// packages/figma-parser/src/full-figma-types.d.ts
export interface Global {
  /** a string uniquely identifying this node within the document */
  readonly id: string;
  /** the name given to the node by the user in the tool. */
  readonly name: string;
  /** whether or not the node is visible on the canvas */
  readonly visible?: boolean;
  /** the type of the node, refer to table below for details */
  readonly type: NodeType;
  /** data written by plugins that is visible only to the plugin that wrote it. Requires the `pluginData` to include the ID of the plugin. */
  readonly pluginData?: any;
  /** data written by plugins that is visible to all plugins. Requires the `pluginData` parameter to include the string "shared". */
  readonly sharedPluginData?: any;
}
/**
 * Styles can be one of the following types
 */
export declare type StyleType = 'FILL' | 'TEXT' | 'EFFECT' | 'GRID';
/**
 * the above styles can be used in the following ways
 */
export declare type StyleKeyType = 'fill' | 'stroke' | 'effect' | 'grid' | 'text' | 'background';
export declare type StylesObject = {
  [key in StyleKeyType]: Record<key, string>;
}[StyleKeyType];
export declare type ScaleMode = 'FILL' | 'FIT' | 'TILE' | 'STRETCH';
export declare type PaintTypeSolid = 'SOLID';
export declare type PaintTypeGraident = 'GRADIENT_LINEAR' | 'GRADIENT_RADIAL' | 'GRADIENT_ANGULAR' | 'GRADIENT_DIAMOND';
export declare type PaintTypeImage = 'IMAGE' | 'EMOJI';
export declare type TextType = 'TEXT';
export declare type PaintType = PaintTypeSolid | PaintTypeGraident | PaintTypeImage;
/**
 * how the layer blends with layers below
 */
export declare type BlendMode = 'PASS_THROUGH' /** (Only applicable to objects with children) */ | 'NORMAL'
  /** Darken: */
  | 'DARKEN' | 'MULTIPLY' | 'LINEAR_BURN' | 'COLOR_BURN'
  /** Lighten: */
  | 'LIGHTEN' | 'SCREEN' | 'LINEAR_DODGE' | 'COLOR_DODGE'
  /** Contrast: */
  | 'OVERLAY' | 'SOFT_LIGHT' | 'HARD_LIGHT'
  /** Inversion: */
  | 'DIFFERENCE' | 'EXCLUSION'
  /** Component: */
  | 'HUE' | 'SATURATION' | 'COLOR' | 'LUMINOSITY';
export declare type EasingType = 'EASE_IN' /** Ease in with an animation curve similar to CSS ease-in */ | 'EASE_OUT' /** Ease out with an animation curve similar to CSS ease-out */ | 'EASE_IN_AND_OUT'; /** Ease in and then out with an animation curve similar to CSS ease-in-out */
export declare type RoleType = 'viewer' | 'editor' | 'owner';
export declare type NodeType = 'DOCUMENT' | 'CANVAS' | 'FRAME' | 'GROUP' | 'VECTOR' | 'BOOLEAN_OPERATION' | 'STAR' | 'LINE' | 'ELLIPSE' | 'REGULAR_POLYGON' | 'RECTANGLE' | 'TEXT' | 'SLICE' | 'COMPONENT' | 'COMPONENT_SET' | 'INSTANCE';
export declare type Node = Document | Canvas | Frame | Group | Vector | BooleanOperation | Star | Line | Ellipse | RegularPolygon | Rectangle | Text | Slice | Component | ComponentSet | Instance;
/** Node Properties */
/** The root node */
export interface Document extends Global {
  readonly type: 'DOCUMENT';
  /** An array of canvases attached to the document */
  readonly children: ReadonlyArray<Node>;
}
/** Represents a single page */
export interface Canvas extends Global {
  readonly type: 'CANVAS';
  /** An array of top level layers on the canvas */
  readonly children: ReadonlyArray<Node>;
  /** Background color of the canvas */
  readonly backgroundColor: Color;
  /** Node ID that corresponds to the start frame for prototypes */
  readonly prototypeStartNodeID: string | null;
  /** An array of export settings representing images to export from the canvas */
  readonly exportSettings?: ReadonlyArray<ExportSetting>;
}
// ...
export interface Frame extends FrameBase {
  readonly type: 'FRAME';
}
/** A logical grouping of nodes */
export interface Group extends FrameBase {
  readonly type: 'GROUP';
}
export interface VectorBase extends Global {
  /**
   * An array of export settings representing images to export from node
   * @default []
   */
  readonly exportSettings?: ReadonlyArray<ExportSetting>;
  /**
   * How this node blends with nodes behind it in the scene
   * (see blend mode section for more details)
   */
  readonly blendMode: BlendMode;
  /**
   * Keep height and width constrained to same ratio
   * @default false
   */
  readonly preserveRatio?: boolean;
  /**
   * Horizontal and vertical layout constraints for node
   */
  readonly constraints: LayoutConstraint;
  /**
   * Node ID of node to transition to in prototyping
   * @default null
   */
  readonly transitionNodeID?: string | null;
  /**
   * The duration of the prototyping transition on this node (in milliseconds)
   * @default null
   */
  readonly transitionDuration?: number | null;
  /**
   * The easing curve used in the prototyping transition on this node
   * @default null
   */
  readonly transitionEasing?: EasingType | null;
  /**
   * Opacity of the node
   * @default 1
   */
  readonly opacity?: number;
  /** Bounding box of the node in absolute space coordinates */
  readonly absoluteBoundingBox: Rect;
  /**
   * Width and height of element. This is different from the width and height
   * of the bounding box in that the absolute bounding box represents the
   * element after scaling and rotation. Only present if geometry=paths
   * is passed
   */
  readonly size?: Vector2;
  /**
   * The top two rows of a matrix that represents the 2D transform of this
   * node relative to its parent. The bottom row of the matrix is implicitly
   * always (0, 0, 1). Use to transform coordinates in geometry.
   * Only present if geometry=paths is passed
   */
  readonly relativeTransform?: Transform;
  /**
   * An array of effects attached to this node
   * (see effects sectionfor more details)
   * @default []
   */
  readonly effects: ReadonlyArray<Effect>;
  /**
   * Does this node mask sibling nodes in front of it?
   * @default false
   */
  readonly isMask?: boolean;
  /**
   * An array of fill paints applied to the node
   * @default []
   */
  readonly fills: ReadonlyArray<Paint>;
  /**
   * Only specified if parameter geometry=paths is used. An array of paths
   * representing the object fill
   */
  readonly fillGeometry?: ReadonlyArray<Path>;
  /**
   * An array of stroke paints applied to the node
   * @default []
   */
  readonly strokes: ReadonlyArray<Paint>;
  /** The weight of strokes on the node */
  readonly strokeWeight: number;
  /**
   * Only specified if parameter geometry=paths is used. An array of paths
   * representing the object stroke
   */
  readonly strokeGeometry?: ReadonlyArray<Path>;
  /**
   * Where stroke is drawn relative to the vector outline as a string enum
   * "INSIDE": draw stroke inside the shape boundary
   * "OUTSIDE": draw stroke outside the shape boundary
   * "CENTER": draw stroke centered along the shape boundary
   */
  readonly strokeAlign: 'INSIDE' | 'OUTSIDE' | 'CENTER';
  /**
   * Styles this node uses from the global `styles`
   */
  readonly styles?: StylesObject;
}
/** A vector network, consisting of vertices and edges */
export interface Vector extends VectorBase {
  readonly type: 'VECTOR';
}
/** A group that has a boolean operation applied to it */
export interface BooleanOperation extends VectorBase {
  readonly type: 'BOOLEAN_OPERATION';
  /**
   * A string enum with value of "UNION", "INTERSECT", "SUBTRACT", or "EXCLUDE"
   * indicating the type of boolean operation applied
   */
  readonly booleanOperation: 'UNION' | 'INTERSECT' | 'SUBTRACT' | 'EXCLUDE';
  /** An array of nodes that are being boolean operated on */
  readonly children: ReadonlyArray<Node>;
}
/** A regular star shape */
export interface Star extends VectorBase {
  readonly type: 'STAR';
}
/** A straight line */
export interface Line extends VectorBase {
  readonly type: 'LINE';
}
/** An ellipse */
export interface Ellipse extends VectorBase {
  readonly type: 'ELLIPSE';
}
/** A regular n-sided polygon */
export interface RegularPolygon extends VectorBase {
  readonly type: 'REGULAR_POLYGON';
}
/** A rectangle */
export interface Rectangle extends VectorBase {
  readonly type: 'RECTANGLE';
  /** Radius of each corner of the rectangle if a single radius is set for all corners */
  readonly cornerRadius?: number;
  /** Array of length 4 of the radius of each corner of the rectangle, starting in the top left and proceeding clockwise */
  readonly rectangleCornerRadii?: readonly [number, number, number, number];
}
/** A text box */
export interface Text extends VectorBase {
  readonly type: TextType;
  /** Text contained within text box */
  readonly characters: string;
  /**
   * Style of text including font family and weight (see type style
   * section for more information)
   */
  readonly style: TypeStyle;
  /**
   * Array with same number of elements as characeters in text 'box' |    * each element is a reference to the styleOverrideTable defined
   * below and maps to the corresponding character in the characters
   * field. Elements with value 0 have the default type style
   */
  readonly characterStyleOverrides: ReadonlyArray<number>;
  /** Map from ID to TypeStyle for looking up style overrides */
  readonly styleOverrideTable: {
    readonly [index: number]: TypeStyle;
  };
}
/** A rectangular region of the canvas that can be exported */
export interface Slice extends Global {
  readonly type: 'SLICE';
  /** An array of export settings representing images to export from this node */
  readonly exportSettings: ReadonlyArray<ExportSetting>;
  /** Bounding box of the node in absolute space coordinates */
  readonly absoluteBoundingBox: Rect;
  /**
   * Width and height of element. This is different from the width and height
   * of the bounding box in that the absolute bounding box represents the
   * element after scaling and rotation. Only present if geometry=paths
   * is passed
   */
  readonly size?: Vector2;
  /**
   * The top two rows of a matrix that represents the 2D transform of this
   * node relative to its parent. The bottom row of the matrix is implicitly
   * always (0, 0, 1). Use to transform coordinates in geometry.
   * Only present if geometry=paths is passed
   */
  readonly relativeTransform?: Transform;
}
/** A node that can have instances created of it that share the same properties */
export interface Component extends FrameBase {
  readonly type: 'COMPONENT';
}
/** A node that can have multiple component variations */
export interface ComponentSet extends FrameBase {
  readonly type: 'COMPONENT_SET';
}
/**
 * An instance of a component, changes to the component result in the same
 * changes applied to the instance
 */
export interface Instance extends FrameBase {
  readonly type: 'INSTANCE';
  /**
   * ID of component that this instance came from, refers to components
   * table (see endpoints section below)
   */
  readonly componentId: string;
}
/** An RGBA color */
export interface Color {
  /** Red channel value, between 0 and 1 */
  readonly r: number;
  /** Green channel value, between 0 and 1 */
  readonly g: number;
  /** Blue channel value, between 0 and 1 */
  readonly b: number;
  /** Alpha channel value, between 0 and 1 */
  readonly a: number;
}
/** Format and size to export an asset at */
export interface ExportSetting {
  /** File suffix to append to all filenames */
  readonly suffix: string;
  /** Image type, string enum */
  readonly format: 'JPG' | 'PNG' | 'SVG' | 'PDF';
  /** Constraint that determines sizing of exported asset */
  readonly constraint: Constraint;
}
// ...
export interface Rect {
  /** X coordinate of top left corner of the rectangle */
  readonly x: number;
  /** Y coordinate of top left corner of the rectangle */
  readonly y: number;
  /** Width of the rectangle */
  readonly width: number;
  /** Height of the rectangle */
  readonly height: number;
}
// ...
export declare type Transform = ReadonlyArray<ReadonlyArray<number>>;
/** A 2d vector */
export interface Vector2 {
  /** X coordinate of the vector */
  readonly x: number;
  /** Y coordinate of the vector */
  readonly y: number;
}
/** A position color pair representing a gradient stop */
export interface ColorStop {
  /** Value between 0 and 1 representing position along gradient axis */
  readonly position: number;
  /** Color attached to corresponding position */
  readonly color: Color;
}
/** Metadata for character formatting */
export interface TypeStyle {
  /** Font family of text (standard name) */
  readonly fontFamily: string;
  /** PostScript font name */
  readonly fontPostScriptName: string;
  /** Space between paragraphs in px, 0 if not present */
  readonly paragraphSpacing?: number;
  /** Paragraph indentation in px, 0 if not present */
  readonly paragraphIndent?: number;
  /** Is text italicized? */
  readonly italic?: boolean;
  /** Numeric font weight */
  readonly fontWeight: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
  /** Font size in px */
  readonly fontSize: number;
  /** Horizontal text alignment as string enum */
  readonly textAlignHorizontal: 'LEFT' | 'RIGHT' | 'CENTER' | 'JUSTIFIED';
  /** Vertical text alignment as string enum */
  readonly textAlignVertical: 'TOP' | 'CENTER' | 'BOTTOM';
  /** Space between characters in px */
  readonly letterSpacing: number;
  /** Paints applied to characters */
  readonly fills?: ReadonlyArray<Paint>;
  /** Line height in px */
  readonly lineHeightPx: number;
  /** Line height as a percentage of normal line height */
  readonly lineHeightPercent: number;
  /** The unit of the line height value specified by the user. */
  readonly lineHeightUnit: 'PIXELS' | 'FONT_SIZE_%' | 'INTRINSIC_%';
  /** Text casing applied to the node, default is the original casing */
  readonly textCase?: 'UPPER' | 'LOWER' | 'TITLE';
  /** Text decoration applied to the node, default is none */
  readonly textDecoration?: 'STRIKETHROUGH' | 'UNDERLINE';
  /** Line height as a percentage of the font size. Only returned when lineHeightPercent is not 100. */
  readonly lineHeightPercentFontSize?: number;
}